add "ConnectPassthrough" method to pgconn package
For the proxy to work, we need to be able to re-use jackc/pgx.PgConn,
but also be able to authenticate by passing the packets from the client
app all the way up to the server. To do that, we need this fork and this
patch to make that happen.